feat: add endpoint for FHIR bundle status check, statusURL, and interaction view #35
Added a new endpoint /Bundle/$status/{bundleSessionId} in Controller.java to get the status of the FHIR bundle validate and create process by passing an ID.
Included a statusURL element in the immediateResult variable, generating a URL with request.getServerName() and bundleAsyncInteractionId.toString() for easy access to the status.
Added proper interaction view to improve the visibility and handling of interactions within the FHIR bundle process.